Backend/pre_processing/requirements.txt
paddleocr==3.4.0
paddlepaddle==3.0.0
paddlex==3.4.3
opencv-contrib-python==4.10.0.84

pillow==12.2.0
reportlab==4.4.10
PyMuPDF==1.27.2.2
pypdfium2==5.6.0
pdf2image==1.17.0

numpy==2.2.6
scipy==1.15.3
scikit-learn==1.7.2
shapely==2.1.2
networkx==3.4.2

pandas==2.3.3
PyYAML==6.0.2
requests==2.33.1
tqdm==4.67.3
joblib==1.5.3


fastapi 
uvicorn[standard] 
python-multipart
python-dotenv>=1.0.1
azure-storage-blob>=12.0.0